+++ src/components/Layout/LayoutVertical/resourceV2.tsx
@@ -290,18 +290,17 @@ export const resource: ItemType[] =[
     ],
     icon: <i className="fa-solid fa-code-branch"></i>,
   },
-
   // Báo cáo
   {
     label: "Báo cáo",
     key: "report",
-    permission: [POLICIES.READ_REPORTSALARY],//
+    permission: [POLICIES.READ_REPORTSALARY],
     children: [
       {
-        label: "Báo cáo nhân viên",
+        label: "Báo cáo lương trình dược viên",
         path: PATH_APP.report.employee,
         key: PATH_APP.report.employee,
-        permission: [POLICIES.READ_REPORTSALARY],//
+        permission : [POLICIES.READ_REPORTSALARY]
       },
     ],
     icon: <i className="fa-solid fa-code-branch"></i>,