+++ src/modules/salesGroup/components/Action.tsx
@@ -1,18 +1,16 @@
 import {
-    ApartmentOutlined,
-    DeleteOutlined,
-    InfoCircleOutlined,
-    MoreOutlined,
-    SisternodeOutlined,
+  ApartmentOutlined,
+  DeleteOutlined,
+  InfoCircleOutlined,
+  MoreOutlined,
+  SisternodeOutlined
 } from "@ant-design/icons";
 import { Button, Dropdown, MenuProps, Popconfirm, Tooltip } from "antd";
 import { get } from "lodash";
 import WithPermission from "~/components/common/WithPermission";
 import POLICIES from "~/modules/policy/policy.auth";
-import { SALES_GROUP_GEOGRAPHY } from "../constants";
-import { RulesLeader, RulesMember } from "../salesGroup.service";
+import { RulesLeader } from "../salesGroup.service";
 import useSalesGroupStore from "../salesGroupContext";
-const RulesMemberMethod = new RulesMember();
 const RulesLeaderMethod = new RulesLeader();
 type PropsType = {
   _id: string;
@@ -29,32 +27,27 @@ const Action = ({ _id, rc }: PropsType) => {
     canWrite,
     canUpdate,
   } = useSalesGroupStore();
-  const existMember = RulesMemberMethod.isExist(
-    get(rc, "salesGroupPermission", [])
-  );
   const notExistLeader = !RulesLeaderMethod.isExist(
     get(rc, "salesGroupPermission", [])
   );
-  const isZone = get(rc, "typeArea") === SALES_GROUP_GEOGRAPHY.ZONE;
 
   const items: MenuProps["items"] = [
     {
       label: (
         <Tooltip
           title={
-            isZone || existMember || notExistLeader
+              notExistLeader
               ? "Đã có TDV hoặc đã là nhóm nhỏ hoặc không có quản lý"
               : ""
           }
         >
         {canWrite &&  <Button
-            disabled={existMember || isZone || notExistLeader}
+            disabled={ notExistLeader}
             block
             icon={<SisternodeOutlined />}
             onClick={() =>
               onOpenFormCreateGroupFromExistGroup({
                 parentNear: _id,
-                parentNearName: get(rc, "name", ""),
                 parentNearPath: get(rc, "managementArea", [])?.map(
                   (area: any) => get(area, "path")
                 ),