+++ src/modules/salesGroup/components/Relationship/CardRelation.tsx
@@ -1,10 +1,13 @@
-import { Badge, Flex, Tag, Typography } from "antd";
+import { Flex, Tag } from "antd";
 import { get } from "lodash";
-import React from "react";
+import React, { useMemo } from "react";
 import WhiteBox from "~/components/common/WhiteBox";
-import {  SALES_GROUP_GEOGRAPHY_COLOR,SALES_GROUP_GEOGRAPHY_VI } from "../../constants";
+import { EMPLOYEE_LEVEL, EMPLOYEE_LEVEL_COLOR } from "~/modules/employee/constants";
 import Address from "../Address";
+import ListMember from "../ListMember";
 import PopoverCardEmployee from "../PopoverCardEmployee";
+const CLONE_EMPLOYEE_LEVEL : any = EMPLOYEE_LEVEL;
+const CLONE_EMPLOYEE_LEVEL_COLOR : any = EMPLOYEE_LEVEL_COLOR;
 type propsType = {
   managementArea?: any[];
   leader?: any;
@@ -14,8 +17,6 @@ type propsType = {
   typeArea?: any;
 };
 
-const CLONE_SALES_GROUP_GEOGRAPHY_COLOR : any = SALES_GROUP_GEOGRAPHY_COLOR;
-const CLONE_SALES_GROUP_GEOGRAPHY_VI : any = SALES_GROUP_GEOGRAPHY_VI;
 export default function CardRelation({
   managementArea,
   leader,
@@ -24,29 +25,27 @@ export default function CardRelation({
   name,
   typeArea,
 }: propsType): React.JSX.Element {
+  const employeeLevelLeader = useMemo(() => get(leader, "employee.employeeLevel", ""),[leader])
   return (
-  <Tag style={{padding : 5}} color={CLONE_SALES_GROUP_GEOGRAPHY_COLOR[typeArea]}>
-      <Flex align={"center"} vertical>
-      {/* <WhiteBox> */}
+      <Flex align={"center"} vertical >
+      <WhiteBox>
         <Flex align={"center"} vertical>
-        <Typography.Text strong><Tag color={CLONE_SALES_GROUP_GEOGRAPHY_COLOR[typeArea]}>{CLONE_SALES_GROUP_GEOGRAPHY_VI[typeArea]}</Tag>{name}</Typography.Text>
         
         {leader && (
           <div>
             Quản lý:{" "}
             <PopoverCardEmployee employee={get(leader, "employee", "")}>
-              <span style={{cursor: 'default'}}>{get(leader, "employee.fullName", "")}</span>
+              <Tag color={CLONE_EMPLOYEE_LEVEL_COLOR[employeeLevelLeader]}>{CLONE_EMPLOYEE_LEVEL[employeeLevelLeader]}</Tag><span style={{cursor: 'default'}}>{get(leader, "employee.fullName", "")}</span>
             </PopoverCardEmployee>
           </div>
         )}
-        {member && (
-          <div>
-            TDV:{" "}
-            <PopoverCardEmployee employee={get(member, "employee", "")}>
+          <Flex align={'center'} gap={5}>
+            Thành viên:{" "}
+            {/* <PopoverCardEmployee employee={get(member, "employee", "")}>
               <span style={{cursor: 'default'}}>{get(member, "employee.fullName", "")}</span>
-            </PopoverCardEmployee>
-          </div>
-        )}
+            </PopoverCardEmployee> */}
+            {member?.length ? <ListMember member={member}/> : "Trống"}
+          </Flex>
         {!leader && !member ? <Tag bordered={false} color={'error'}>Trống</Tag> : null}
         <Address
           onlyShowLastPath={!!parentNear}
@@ -54,9 +53,9 @@ export default function CardRelation({
           bordered={true}
         />
         </Flex>
-      {/* </WhiteBox> */}
+      </WhiteBox>
       
     </Flex>
-  </Tag>
+  // </Tag>
   );
 }