+++ src/modules/salesGroup/screens/SalesGroup.tsx
@@ -1,4 +1,5 @@
-import { Button, Tag } from "antd";
+import { ApartmentOutlined, AppstoreOutlined } from "@ant-design/icons";
+import { Button, Flex } from "antd";
 import { ColumnsType } from "antd/es/table/InternalTable";
 import { get } from "lodash";
 import { useRef, useState } from "react";
@@ -7,17 +8,16 @@ import TableAnt from "~/components/Antd/TableAnt";
 import Breadcrumb from "~/components/common/Breadcrumb";
 import SelectSearch from "~/components/common/SelectSearch/SelectSearch";
 import WhiteBox from "~/components/common/WhiteBox";
+import POLICIES from "~/modules/policy/policy.auth";
 import { StringToSlug } from "~/utils/helpers";
 import Action from "../components/Action";
 import Address from "../components/Address";
+import ExchangeRate from "../components/ExchangeRate";
 import Member from "../components/Member";
 import Relationship from "../components/Relationship";
 import SalesGroupForm from "../components/SalesGroupForm";
+import SalesGroupTree from "../components/SalesGroupTree";
 import TargetSalesGroup from "../components/TargetSalesGroup";
-import {
-  SALES_GROUP_GEOGRAPHY_COLOR,
-  SALES_GROUP_GEOGRAPHY_VI
-} from "../constants";
 import {
   useGetSalesGroups,
   useGetSalesGroupsSearch,
@@ -25,11 +25,6 @@ import {
 } from "../salesGroup.hook";
 import { SalesGroupType } from "../salesGroup.modal";
 import useSalesGroupStore from "../salesGroupContext";
-import ExchangeRate from "../components/ExchangeRate";
-import TableSelect from "../components/ExchangeRate/TableSelect";
-import POLICIES from "~/modules/policy/policy.auth";
-const CLONE_SALES_GROUP_GEOGRAPHY_VI: any = SALES_GROUP_GEOGRAPHY_VI;
-const CLONE_SALES_GROUP_GEOGRAPHY_COLOR: any = SALES_GROUP_GEOGRAPHY_COLOR;
 export default function SalesGroup() {
   const {
     updateSalesGroup,
@@ -54,7 +49,8 @@ export default function SalesGroup() {
   const [expandedRowKeys, setExpandedRowKeys]: any = useState([]);
   const [query] = useSalesGroupQueryParams();
   const [data, isLoading, actionUpdate] = useGetSalesGroups(query);
-
+  const [modeView,setModeView] = useState<'table' | 'tree'>('table');
+  
   const dataSearch = useGetSalesGroupsSearch();
 
   const onSearch = (keyword: any) => {
@@ -120,18 +116,18 @@ export default function SalesGroup() {
         );
       },
     },
-    {
-      title: "Loại nhóm",
-      key: "typeArea",
-      dataIndex: "typeArea",
-      width: 140,
-      align: "center",
-      render: (typeArea, rc) => (
-        <Tag color={CLONE_SALES_GROUP_GEOGRAPHY_COLOR[typeArea]}>
-          {CLONE_SALES_GROUP_GEOGRAPHY_VI?.[typeArea]}
-        </Tag>
-      ),
-    },
+    // {
+    //   title: "Loại nhóm",
+    //   key: "typeArea",
+    //   dataIndex: "typeArea",
+    //   width: 140,
+    //   align: "center",
+    //   render: (typeArea, rc) => (
+    //     <Tag color={CLONE_SALES_GROUP_GEOGRAPHY_COLOR[typeArea]}>
+    //       {CLONE_SALES_GROUP_GEOGRAPHY_VI?.[typeArea]}
+    //     </Tag>
+    //   ),
+    // },
     {
       title: "Chỉ tiêu",
       key: "targets",
@@ -196,7 +192,13 @@ export default function SalesGroup() {
           isShowButtonAdd
           permissionKey={[POLICIES.WRITE_SALESGROUP]}
         />
-        <TableAnt
+        <Flex style={{marginTop : 20}}>
+          <Button.Group>
+            <Button onClick={() => setModeView('table')} type={modeView === 'table' ? 'primary' : 'default'} icon={<AppstoreOutlined />}/>
+            <Button onClick={() => setModeView('tree')} type={modeView === 'tree' ? 'primary' : 'default'} icon={<ApartmentOutlined />}/>
+          </Button.Group>
+        </Flex>
+        {modeView === 'table' && <TableAnt
           expandable={{
             expandedRowKeys,
             
@@ -224,9 +226,8 @@ export default function SalesGroup() {
           pagination={false}
           bordered
           style={{ marginTop: 20 }}
-          scroll={{ x: 1500 }}
-          stickyTop
-        />
+        />}
+        {modeView === 'tree' && <SalesGroupTree dataSource={dataSearch?.length ? dataSearch : data}/>}
       </WhiteBox>
       <ModalAnt
         onCancel={onCloseForm}
@@ -239,7 +240,6 @@ export default function SalesGroup() {
           onUpdate={updateSalesGroup}
           id={id}
           parentNear={get(parentNear, "parentNear")}
-          parentNearName={get(parentNear, "parentNearName")}
           parentNearPath={get(parentNear, "parentNearPath")}
         />
       </ModalAnt>
@@ -250,7 +250,6 @@ export default function SalesGroup() {
         destroyOnClose
         width={"max-content"}
         centered
-        title="Chi tiết sơ đồ"
       >
         <Relationship id={id} />
       </ModalAnt>